Servo/Sensor is by default 90 degrees to the front
In static tests, when an object is in front, the sensor sweeps the forward 180 degrees at 15 degrees increment

If there is an obstruction in front within 50 cm, it stops.

It then performs a sweep again and tries turning towards the longest distance found.

If it finds an opening, it initiates a turn.